todo/requirements.txt
Django>=4.2
celery>=5.3
djangorestframework>=3.14
djoser>=2.2